Clean up code
- Separate modules
