+++ jira.go
@@ -1,104 +1,16 @@
 package gojira
 
 import (
-	"encoding/json"
-	"encoding/xml"
 	"fmt"
 	"io"
 	"io/ioutil"
 	"net/http"
-	"net/url"
-	"strconv"
-	"time"
 	"math"
 )
 
 const (
 	dateLayout = "2006-01-02T15:04:05.000-0700"
 )
-
-type Jira struct {
-	BaseUrl      string
-	ApiPath      string
-	ActivityPath string
-	GreenHopper	 string
-	Client       *http.Client
-	Auth         *Auth
-}
-
-type Auth struct {
-	Login    string
-	Password string
-}
-
-type Pagination struct {
-	Total      int
-	StartAt    int
-	MaxResults int
-	Page       int
-	PageCount  int
-	Pages      []int
-}
-
-func (p *Pagination) Compute() {
-	p.PageCount = int(math.Ceil(float64(p.Total) / float64(p.MaxResults)))
-	p.Page = int(math.Ceil(float64(p.StartAt) / float64(p.MaxResults)))
-
-	p.Pages = make([]int, p.PageCount)
-	for i := range p.Pages {
-		p.Pages[i] = i
-	}
-}
-
-type JiraProject struct {
-    Self       string `json:"-"`
-	Id         string `json:"-"`
-	Key        string `json:"key"`
-	Name       string `json:"-"`
-	AvatarUrls map[string]string `json:"-"`
-}
-
-type ActivityItem struct {
-	Title    string    `xml:"title"json:"title"`
-	Id       string    `xml:"id"json:"id"`
-	Link     []Link    `xml:"link"json:"link"`
-	Updated  time.Time `xml:"updated"json:"updated"`
-	Author   Person    `xml:"author"json:"author"`
-	Summary  Text      `xml:"summary"json:"summary"`
-	Category Category  `xml:"category"json:"category"`
-}
-
-type ActivityFeed struct {
-	XMLName  xml.Name        `xml:"http://www.w3.org/2005/Atom feed"json:"xml_name"`
-	Title    string          `xml:"title"json:"title"`
-	Id       string          `xml:"id"json:"id"`
-	Link     []Link          `xml:"link"json:"link"`
-	Updated  time.Time       `xml:"updated,attr"json:"updated"`
-	Author   Person          `xml:"author"json:"author"`
-	Entries  []*ActivityItem `xml:"entry"json:"entries"`
-}
-
-
-type Category struct {
-	Term string `xml:"term,attr"json:"term"`
-}
-
-type Link struct {
-	Rel  string `xml:"rel,attr,omitempty"json:"rel"`
-	Href string `xml:"href,attr"json:"href"`
-}
-
-type Person struct {
-	Name     string `xml:"name"json:"name"`
-	URI      string `xml:"uri"json:"uri"`
-	Email    string `xml:"email"json:"email"`
-	InnerXML string `xml:",innerxml"json:"inner_xml"`
-}
-
-type Text struct {
-	Type string `xml:"type,attr,omitempty"json:"type"`
-	Body string `xml:",chardata"json:"body"`
-}
 
 func NewJira(baseUrl string, apiPath string, activityPath string, auth *Auth) *Jira {
 
@@ -124,6 +36,7 @@ func NewJIRA(baseUrl string, auth *Auth) *Jira {
 		GreenHopper:  "/rest/greenhopper/latest",
 		Client:       client,
 		Auth:         auth,
+		Encoding:	  "json",
 	}
 }
 
@@ -149,66 +62,45 @@ func (j *Jira) buildAndExecRequest(method string, url string, data io.Reader) []
 	return contents
 }
 
-func (j *Jira) UserActivity(user string) (ActivityFeed, error) {
-	url := j.BaseUrl + j.ActivityPath + "?streams=" + url.QueryEscape("user IS " + user)
+func (p *Pagination) Compute() {
+	p.PageCount = int(math.Ceil(float64(p.Total) / float64(p.MaxResults)))
+	p.Page = int(math.Ceil(float64(p.StartAt) / float64(p.MaxResults)))
 
-	return j.Activity(url)
+	p.Pages = make([]int, p.PageCount)
+	for i := range p.Pages {
+		p.Pages[i] = i
+	}
 }
 
-func (j *Jira) Activity(url string) (ActivityFeed, error) {
+type Jira struct {
+	BaseUrl      string
+	ApiPath      string
+	ActivityPath string
+	GreenHopper	 string
+	Encoding	 string
 
-	contents := j.buildAndExecRequest("GET", url, nil)
-
-	var activity ActivityFeed
-	err := xml.Unmarshal(contents, &activity)
-	if err != nil {
-		fmt.Println("%s", err)
-	}
-
-	return activity, err
+	Auth         *Auth
+	Client       *http.Client
 }
 
-// search issues assigned to given user
-func (j *Jira) IssuesAssignedTo(user string, maxResults int, startAt int) IssueList {
-
-	url := j.BaseUrl + j.ApiPath + "/search?jql=assignee=\"" + url.QueryEscape(user) + "\"&startAt=" + strconv.Itoa(startAt) + "&maxResults=" + strconv.Itoa(maxResults)
-	contents := j.buildAndExecRequest("GET", url, nil)
-
-	var issues IssueList
-	err := json.Unmarshal(contents, &issues)
-	if err != nil {
-		fmt.Println("%s", err)
-	}
-
-	for _, issue := range issues.Issues {
-        t, _ := time.Parse(dateLayout, issue.Fields.Created)
-        issue.CreatedAt = t
-	}
-
-	pagination := Pagination{
-		Total:      issues.Total,
-		StartAt:    issues.StartAt,
-		MaxResults: issues.MaxResults,
-	}
-	pagination.Compute()
-
-	issues.Pagination = &pagination
-
-	return issues
+type Auth struct {
+	Login    string
+	Password string
 }
 
-// search an issue by its id
-func (j *Jira) Issue(id string) Issue {
-
-	url := j.BaseUrl + j.ApiPath + "/issue/" + id
-	contents := j.buildAndExecRequest("GET", url, nil)
-
-	var issue Issue
-	err := json.Unmarshal(contents, &issue)
-	if err != nil {
-		fmt.Println("%s", err)
-	}
-
-	return issue
+type Pagination struct {
+	Total      int
+	StartAt    int
+	MaxResults int
+	Page       int
+	PageCount  int
+	Pages      []int
 }
 
+type JiraProject struct {
+	Self       string `json:"-"`
+	Id         string `json:"-"`
+	Key        string `json:"key"`
+	Name       string `json:"-"`
+	AvatarUrls map[string]string `json:"-"`
+}
